Consolidate store props events into PropsEvents.ts; remove dead tee-up events and propsGroup
- PropsEvents.ts is now the single source of truth for all store-props
  event types, interfaces, and BatchProcessingStatus enum
- InteractionEvents.ts re-exports from PropsEvents.ts for backward compat
- Removed dead events: SetupStarted, AtmosphericRequest, GameBoxSpawned,
  RendererReady (no live subscribers after coordinator + two-phase refactor)
- Removed SetupStarted emit from StorePropsCoordinator
- Removed RendererReady emit from InstancedShelfRenderer
- Removed StorePropsCoordinator.propsGroup: nothing ever populated it;
  all GPU instanced renderers manage their own scene membership
- Updated unit tests that referenced removed events
