+++ client/test/unit/scene/GpuStorePropsRenderer-initialization.test.ts
@@ -1,205 +1,13 @@
-/**
- * Unit tests for GpuStorePropsRenderer initialization handling
- * 
- * Tests specifically for event-driven initialization waiting (Phase 3 refactoring)
- */
-
-import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest'
-import { EventManager } from '../../../src/core/EventManager'
+import { describe, expect, it } from 'vitest'
 import { StorePropsEventTypes } from '../../../src/types/InteractionEvents'
-import type { RendererReadyEvent } from '../../../src/types/InteractionEvents'
 
-describe('GpuStorePropsRenderer Initialization', () => {
-    let eventManager: EventManager
+describe('StoreProps event surface cleanup', () => {
+    it('does not expose dead tee-up events', () => {
+        const eventTypes = StorePropsEventTypes as Record<string, string | undefined>
 
-    beforeEach(() => {
-        eventManager = EventManager.getInstance()
-    })
-
-    afterEach(() => {
-        vi.clearAllMocks()
-    })
-
-    describe('RendererReady Event Handling', () => {
-        it('should register listener for RendererReady event', () => {
-            const registerSpy = vi.spyOn(eventManager, 'registerEventHandler')
-            
-            // This would be in GpuStorePropsRenderer.setupEventListeners()
-            eventManager.registerEventHandler(
-                StorePropsEventTypes.RendererReady,
-                () => {}
-            )
-            
-            expect(registerSpy).toHaveBeenCalledWith(
-                StorePropsEventTypes.RendererReady,
-                expect.any(Function)
-            )
-        })
-
-        it('should execute queued callbacks when RendererReady event fires', async () => {
-            const initializationQueue: Array<() => void> = []
-            let isShelfRendererReady = false
-            
-            // Simulate GpuStorePropsRenderer's event handler
-            const handleRendererReady = (event: CustomEvent<RendererReadyEvent>) => {
-                if (event.detail.rendererType !== 'shelf') return
-                
-                isShelfRendererReady = true
-                
-                while (initializationQueue.length > 0) {
-                    const callback = initializationQueue.shift()
-                    callback?.()
-                }
-            }
-            
-            eventManager.registerEventHandler(
-                StorePropsEventTypes.RendererReady,
-                handleRendererReady
-            )
-            
-            // Queue some callbacks before event fires
-            let callback1Executed = false
-            let callback2Executed = false
-            
-            initializationQueue.push(() => { callback1Executed = true })
-            initializationQueue.push(() => { callback2Executed = true })
-            
-            // Emit event
-            eventManager.emit<RendererReadyEvent>(
-                StorePropsEventTypes.RendererReady,
-                { rendererType: 'shelf' }
-            )
-            
-            // Wait for event processing
-            await new Promise(resolve => setTimeout(resolve, 0))
-            
-            expect(isShelfRendererReady).toBe(true)
-            expect(callback1Executed).toBe(true)
-            expect(callback2Executed).toBe(true)
-            expect(initializationQueue).toHaveLength(0)
-        })
-
-        it('should handle fast path when renderer already ready', async () => {
-            let isShelfRendererReady = true  // Already initialized
-            const initializationQueue: Array<() => void> = []
-            
-            // Simulate waitForShelfRendererReady fast path
-            const waitForShelfRendererReady = (): Promise<void> => {
-                if (isShelfRendererReady) {
-                    return Promise.resolve()
-                }
-                
-                return new Promise<void>((resolve) => {
-                    initializationQueue.push(resolve)
-                })
-            }
-            
-            // Should resolve immediately
-            const promise = waitForShelfRendererReady()
-            
-            await expect(promise).resolves.toBeUndefined()
-            expect(initializationQueue).toHaveLength(0)
-        })
-
-        it('should handle slow path when renderer not yet ready', async () => {
-            let isShelfRendererReady = false  // Not initialized yet
-            const initializationQueue: Array<() => void> = []
-            
-            // Simulate waitForShelfRendererReady slow path
-            const waitForShelfRendererReady = (): Promise<void> => {
-                if (isShelfRendererReady) {
-                    return Promise.resolve()
-                }
-                
-                return new Promise<void>((resolve) => {
-                    initializationQueue.push(resolve)
-                })
-            }
-            
-            // Start waiting
-            const waitPromise = waitForShelfRendererReady()
-            
-            // Should be queued, not resolved yet
-            expect(initializationQueue).toHaveLength(1)
-            
-            // Simulate RendererReady event
-            isShelfRendererReady = true
-            while (initializationQueue.length > 0) {
-                const callback = initializationQueue.shift()
-                callback?.()
-            }
-            
-            // Should now resolve
-            await expect(waitPromise).resolves.toBeUndefined()
-        })
-
-        it('should only process shelf renderer ready events', async () => {
-            let shelfCallbackExecuted = false
-            let gameboxCallbackExecuted = false
-            
-            const handleRendererReady = (event: CustomEvent<RendererReadyEvent>) => {
-                if (event.detail.rendererType === 'shelf') {
-                    shelfCallbackExecuted = true
-                } else if (event.detail.rendererType === 'gamebox') {
-                    gameboxCallbackExecuted = true
-                }
-            }
-            
-            eventManager.registerEventHandler(
-                StorePropsEventTypes.RendererReady,
-                handleRendererReady
-            )
-            
-            // Emit gamebox event (should be ignored)
-            eventManager.emit<RendererReadyEvent>(
-                StorePropsEventTypes.RendererReady,
-                { rendererType: 'gamebox' }
-            )
-            
-            await new Promise(resolve => setTimeout(resolve, 0))
-            
-            expect(shelfCallbackExecuted).toBe(false)
-            expect(gameboxCallbackExecuted).toBe(true)
-        })
-
-        // TODO: Test error handling - what if RendererReady never fires? 
-        // Should there be a timeout? Should it throw an error?
-        
-        // TODO: Test cleanup - what happens to queued callbacks if renderer is disposed?
-        // Should they be cleared? Should they error?
-        
-        // TODO: Test race condition - what if multiple batches arrive before RendererReady?
-        // Should they all queue properly? Is there a max queue size?
-    })
-
-    describe('Migration from Polling', () => {
-        // These tests document the old behavior for comparison
-        
-        it('OLD: polling used 50ms intervals', () => {
-            // This is documentary - the old code polled every 50ms
-            const pollingInterval = 50
-            expect(pollingInterval).toBe(50)
-        })
-
-        it('OLD: polling had 10 second timeout', () => {
-            // This is documentary - the old code had a 10s timeout
-            const timeout = 10000
-            expect(timeout).toBe(10000)
-        })
-
-        it('NEW: event-driven has no polling overhead', async () => {
-            // Event-driven approach has zero polling overhead
-            let callbackExecuted = false
-            
-            const promise = new Promise<void>((resolve) => {
-                callbackExecuted = true
-                resolve()
-            })
-            
-            await promise
-            
-            expect(callbackExecuted).toBe(true)
-            // No setTimeout, no polling, instant resolution
-        })
+        expect(eventTypes.RendererReady).toBeUndefined()
+        expect(eventTypes.SetupStarted).toBeUndefined()
+        expect(eventTypes.AtmosphericRequest).toBeUndefined()
+        expect(eventTypes.GameBoxSpawned).toBeUndefined()
     })
 })
